AV-1361: Implement front-end functionality
Implement snippets for the spatial query by overriding the snippets
and the css from the spatial search module.

Implement own snippet for searching the municipalities through a drop down
menu with select2.

Implement javscript which hooks into the search of the spatial module and performs
a search based on the drop down bounding box data.

Refs AV-1361