AV-1361: Get bbox data script
Script for generating the JSON file which is used by the spatial
data functionality. The script uses the GeoPy client for using the
geocoding service. The script uses the Nominatim service by default,
it has not been tested with anything else. Changing service should
be possible by modifying the code.

Refs AV-1361